backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
alembic==1.13.1
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
python-pptx
pydantic_settings
openai
